+++ preprocess_ssl.py
@@ -9,23 +9,26 @@ import utils
 from wavlm import WavLM, WavLMConfig
 
 
-def process(filename):
-    basename = os.path.basename(filename)
-    # speaker = basename[:4]
-    speaker, *_ = basename.split("-")
+def process(file_path):
+    basename = os.path.basename(file_path)
+    dirname = os.path.dirname(file_path)
+    speaker = os.path.dirname(dirname)
+    
     save_dir = os.path.join(args.out_dir, speaker)
     os.makedirs(save_dir, exist_ok=True)
-    wav, _ = librosa.load(filename, sr=args.sr)
-    wav = torch.from_numpy(wav).unsqueeze(0).cuda()
-    c = utils.get_content(cmodel, wav)
+    
     save_name = os.path.join(save_dir, basename.replace(".wav", ".pt"))
-    torch.save(c.cpu(), save_name)
+    if not os.path.exists(save_name):
+        wav, _ = librosa.load(file_path, sr=args.sr)
+        wav = torch.from_numpy(wav).unsqueeze(0).cuda()
+        c = utils.get_content(cmodel, wav)
+        torch.save(c.cpu(), save_name)
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("--sr", type=int, default=16000, help="sampling rate")
-    parser.add_argument("--in_dir", type=str, default="dataset/arknights-jp-16k", help="path to input dir")
+    parser.add_argument("--in_dir", type=str, default="dataset/universal-16k", help="path to input dir")
     parser.add_argument("--out_dir", type=str, default="dataset/wavlm", help="path to output dir")
     args = parser.parse_args()
